Moved around the test exercises and started with the cypress exercise
