--- week2/test-exercises/README.md
@@ -1,11 +1,0 @@
-# Test exercises
-
-During the curriculum we looked at testing a couple of times, but overall we expect that you have avoided the topic of testing a bit ;). That is understandable as testing is very difficult, not very glamorous and is not very useful for the solo, one-time projects you are generally making during the curriculum.
-
-As we are nearing the end of the curriculum it is now time to stop avoiding the topic however and start giving testing the attention it deserves. In this sample project you will find some component files (files that end in `.js`) with associated test files (files that end in `.test.js`). The test files will have comments to help you out, we have also already described the things you should be testing. It is your job to make sure that all of the tests pass.
-
-You may edit the components if you wish, but only make changes that help you test it. The functionality of the component should stay the same!
-
-To set up this folder, please run `npm run install`.
-
-To run the tests you can run `npm test`.
