+++ week3/test-exercises/README.md
@@ -1,17 +1,20 @@
 # Test exercises
 
-This week we will continue practicing testing by adding a couple more scenario's that you will encounter in your real world applications. Last week we looked at the following things:
-
-- A rendering component with data props and no other user interaction
-- A form component with state changes and a button for user interaction
-- A simple component that grabs data from an api and renders it on the screen.
-
-This week we learned about client-side routing, the context API and custom hooks. These are great things to use in your apps but they can complicate things a bit when testing. So let's try to deal with them one by one!
-
 In this sample project you will find some component files (files that end in `.js`) with associated test files (files that end in `.test.js`). The test files will have comments to help you out, we have also already described the things you should be testing. It is your job to make sure that all of the tests pass. You will notice that there are components with no associated test files, you only need to do the ones that have a test file associated with it.
+
+These test exercises try to provide you with a couple scenario's that you will encounter in your real world applications that you may be asked to test. We have the following things:
+
+- `1-UserListItem` - A rendering component with data props and no other user interaction
+- `2-UserDetailsForm` - A form component with state changes and a button for user interaction
+- `3-ChuckNorrisJoke` - A simple component that grabs data from an api and renders it on the screen
+- `4-Routing` - A complete router that switches between pages
+- `5-Context` - A component set up to use a context
+- `6-userListState` - A custom hook that can be tested.
 
 You may edit the components if you wish, but only make changes that help you test it. The functionality of the component should stay the same!
 
 To set up this folder, please run `npm run install`.
 
 To run the tests you can run `npm test`.
+
+Note: this kind of test coverage is quite high and will only be necessary if the company you are going to work in deems it necessary. It is more important to have a good grasp of React than be able to do this, but we leave the tests here anyway for you to practice this kind of testing in an isolated way.
